Lexer: Refactor the whitespace handling in all states
Previously the whitespace tokens were handled for all states with a
global rule. That is not a problem right now, but will make trouble
with the upcoming export statement support.

The export statement has special requirements regarding whitespaces
and therefore some states will be needed that must not be affected by
the global whitespace handling rule.

Remove the global rule and replace it by state specific rules where
necessary.

Signed-off-by: Florian Bezdeka <florian@bezdeka.de>